CORE-19437 Moving DurableProcessor to seperate class
