+++ components/flow/flow-rest-resource-service-impl/src/test/kotlin/net/corda/flow/rest/impl/DurableFlowStatusProcessorTest.kt
@@ -1,161 +1,55 @@
 package net.corda.flow.rest.impl
 
-import net.corda.avro.serialization.CordaAvroSerializationFactory
 import net.corda.avro.serialization.CordaAvroSerializer
 import net.corda.data.flow.FlowInitiatorType
 import net.corda.data.flow.FlowKey
 import net.corda.data.flow.output.FlowStates
 import net.corda.data.flow.output.FlowStatus
-import net.corda.data.identity.HoldingIdentity
-import net.corda.libs.configuration.SmartConfig
 import net.corda.libs.statemanager.api.IntervalFilter
 import net.corda.libs.statemanager.api.MetadataFilter
 import net.corda.libs.statemanager.api.State
 import net.corda.libs.statemanager.api.StateManager
-import net.corda.libs.statemanager.api.StateManagerFactory
 import net.corda.lifecycle.LifecycleCoordinatorName
-import net.corda.lifecycle.LifecycleEventHandler
-import net.corda.lifecycle.LifecycleStatus
-import net.corda.lifecycle.RegistrationHandle
-import net.corda.lifecycle.RegistrationStatusChangeEvent
-import net.corda.lifecycle.StartEvent
 import net.corda.messaging.api.records.Record
-import net.corda.messaging.api.subscription.Subscription
-import net.corda.messaging.api.subscription.config.SubscriptionConfig
-import net.corda.messaging.api.subscription.factory.SubscriptionFactory
 import net.corda.schema.Schemas
-import net.corda.schema.Schemas.Flow.FLOW_STATUS_TOPIC
-import net.corda.schema.configuration.BootConfig.INSTANCE_ID
-import net.corda.schema.configuration.ConfigKeys
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
-import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
-import org.mockito.kotlin.same
-import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import java.time.Instant
 
-class DurableFlowStatusCacheServiceImplTest {
-    private val subscriptionFactory = mock<SubscriptionFactory>()
-    private val lifecycleTestContext = LifecycleTestContext()
-    private val lifecycleCoordinator = lifecycleTestContext.lifecycleCoordinator
-    private val lifecycleEventRegistration = mock<RegistrationHandle>()
-    private val cordaSerializationFactory = mock<CordaAvroSerializationFactory>()
-    private val stateManagerFactory = mock<StateManagerFactory>()
-    private val serializer = mock<CordaAvroSerializer<FlowStatus>>()
-
-    private var eventHandler = mock<LifecycleEventHandler>()
-    private val topicSubscription = mock<Subscription<FlowKey, FlowStatus>>()
-    private val topicSubscriptionName = LifecycleCoordinatorName("Test Flow Status Subscription")
-    private lateinit var flowStatusCacheService: DurableFlowStatusCacheServiceImpl
+class DurableFlowStatusProcessorTest {
+    private lateinit var flowStatusProcessor: DurableFlowStatusProcessor
     private lateinit var stateManager: StateManager
-    private val config = mock<SmartConfig> {
-        whenever(it.getInt(INSTANCE_ID)).thenReturn(2)
-        whenever(it.getConfig(ConfigKeys.STATE_MANAGER_CONFIG)).thenReturn(mock())
-    }
 
-    companion object {
-        val FLOW_KEY_1 = FlowKey("a1", HoldingIdentity("b1", "c1"))
-        val FLOW_KEY_2 = FlowKey("a2", HoldingIdentity("b1", "c1"))
-    }
+    private val serializer = mock<CordaAvroSerializer<Any>>()
 
     @BeforeEach
     fun setup() {
-        whenever(lifecycleCoordinator.followStatusChangesByName(any())).thenReturn(lifecycleEventRegistration)
-
-        whenever(topicSubscription.subscriptionName).thenReturn(topicSubscriptionName)
-        whenever(subscriptionFactory.createDurableSubscription<FlowKey, FlowStatus>(any(), any(), any(), anyOrNull()))
-            .thenReturn(topicSubscription)
+        stateManager = getMockStateManager()
 
         whenever(serializer.serialize(any())).thenReturn("test".toByteArray())
-        whenever(cordaSerializationFactory.createAvroSerializer<FlowStatus>(any())).thenReturn(serializer)
-
-        stateManager = getMockStateManager()
-        whenever(stateManagerFactory.create(any())).thenReturn(stateManager)
-
-        flowStatusCacheService = DurableFlowStatusCacheServiceImpl(
-            subscriptionFactory,
-            lifecycleTestContext.lifecycleCoordinatorFactory,
-            cordaSerializationFactory,
-            stateManagerFactory
-        )
-
-        eventHandler = lifecycleTestContext.getEventHandler()
+        flowStatusProcessor = DurableFlowStatusProcessor(stateManager, serializer)
     }
 
     @Test
-    fun `Test compacted subscription key class is flow status key`() {
-        assertThat(flowStatusCacheService.keyClass).isEqualTo(FlowKey::class.java)
+    fun `Test durable subscription key class is flow status key`() {
+        assertThat(flowStatusProcessor.keyClass).isEqualTo(FlowKey::class.java)
     }
 
     @Test
-    fun `Test compacted subscription value class is flow status`() {
-        assertThat(flowStatusCacheService.valueClass).isEqualTo(FlowStatus::class.java)
-    }
-
-    @Test
-    fun `Test start starts the lifecycle coordinator`() {
-        flowStatusCacheService.start()
-        verify(lifecycleCoordinator).start()
-    }
-
-    @Test
-    fun `Test stop stops the lifecycle coordinator`() {
-        flowStatusCacheService.stop()
-        verify(lifecycleCoordinator).stop()
-    }
-
-    @Test
-    fun `Test initialise creates new topic subscription and starts it`() {
-        flowStatusCacheService.initialise(config)
-
-        val expectedSubscriptionCfg = SubscriptionConfig(
-            "flow_status_subscription",
-            Schemas.Flow.FLOW_STATUS_TOPIC
-        )
-
-        verify(subscriptionFactory).createDurableSubscription(
-            eq(expectedSubscriptionCfg),
-            same(flowStatusCacheService),
-            same(config),
-            same(null)
-        )
-
-        verify(topicSubscription).start()
-    }
-
-    @Test
-    fun `Test initialise closes any existing topic subscription`() {
-        flowStatusCacheService.initialise(config)
-        // second time around we close the existing subscription
-        flowStatusCacheService.initialise(config)
-        verify(topicSubscription).close()
-    }
-
-    @Test
-    fun `Test on start event component status up is signaled`() {
-        eventHandler.processEvent(StartEvent(), lifecycleCoordinator)
-        verify(lifecycleCoordinator).updateStatus(LifecycleStatus.UP)
-    }
-
-    @Test
-    fun `Test on topic subscription going down component status down is signaled`() {
-        flowStatusCacheService.initialise(config)
-        eventHandler.processEvent(RegistrationStatusChangeEvent(lifecycleEventRegistration, LifecycleStatus.DOWN), lifecycleCoordinator)
-        verify(lifecycleCoordinator).updateStatus(LifecycleStatus.DOWN)
+    fun `Test durable subscription value class is flow status`() {
+        assertThat(flowStatusProcessor.valueClass).isEqualTo(FlowStatus::class.java)
     }
 
     @Test
     fun `Test onNext creates a new record in the StateManager`() {
-        val record = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus())
-        val key = FLOW_KEY_1.toString()
+        val record = Record(Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus())
+        val key = FlowStatusLookupServiceImplTest.FLOW_KEY_1.toString()
 
-        flowStatusCacheService.initialise(config)
-        flowStatusCacheService.onNext(listOf(record))
+        flowStatusProcessor.onNext(listOf(record))
 
         val result = stateManager.get(setOf(key))
 
@@ -166,13 +60,12 @@ class DurableFlowStatusCacheServiceImplTest {
 
     @Test
     fun `Test onNext creates multiple records in the StateManager on different keys`() {
-        val record1 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus())
-        val record2 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_2, createFlowStatus())
-        val key1 = FLOW_KEY_1.toString()
-        val key2 = FLOW_KEY_2.toString()
+        val record1 = Record(Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus())
+        val record2 = Record(Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_2, createFlowStatus())
+        val key1 = FlowStatusLookupServiceImplTest.FLOW_KEY_1.toString()
+        val key2 = FlowStatusLookupServiceImplTest.FLOW_KEY_2.toString()
 
-        flowStatusCacheService.initialise(config)
-        flowStatusCacheService.onNext(listOf(record1, record2))
+        flowStatusProcessor.onNext(listOf(record1, record2))
 
         val result = stateManager.get(setOf(key1, key2))
 
@@ -185,12 +78,15 @@ class DurableFlowStatusCacheServiceImplTest {
 
     @Test
     fun `Test onNext updates existing key if it already exists`() {
-        val record1 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus(FlowStates.START_REQUESTED))
-        val record2 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus(FlowStates.RUNNING))
-        val key = FLOW_KEY_1.toString()
+        val record1 = Record(
+            Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus(
+                FlowStates.START_REQUESTED))
+        val record2 = Record(
+            Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus(
+                FlowStates.RUNNING))
+        val key = FlowStatusLookupServiceImplTest.FLOW_KEY_1.toString()
 
-        flowStatusCacheService.initialise(config)
-        flowStatusCacheService.onNext(listOf(record1))
+        flowStatusProcessor.onNext(listOf(record1))
 
         val result1 = stateManager.get(setOf(key))
 
@@ -198,7 +94,7 @@ class DurableFlowStatusCacheServiceImplTest {
         assertThat(result1.containsKey(key)).isTrue()
         assertThat(result1[key]?.version).isEqualTo(0)
 
-        flowStatusCacheService.onNext(listOf(record2))
+        flowStatusProcessor.onNext(listOf(record2))
 
         val result2 = stateManager.get(setOf(key))
 
@@ -209,16 +105,19 @@ class DurableFlowStatusCacheServiceImplTest {
 
     @Test
     fun `Test onNext processes a create and an update in a single call`() {
-        val record1 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus(FlowStates.START_REQUESTED))
-        val record2 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_1, createFlowStatus(FlowStates.RUNNING))
-        val record3 = Record(FLOW_STATUS_TOPIC, FLOW_KEY_2, createFlowStatus())
+        val record1 = Record(
+            Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus(
+                FlowStates.START_REQUESTED))
+        val record2 = Record(
+            Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_1, createFlowStatus(
+                FlowStates.RUNNING))
+        val record3 = Record(Schemas.Flow.FLOW_STATUS_TOPIC, FlowStatusLookupServiceImplTest.FLOW_KEY_2, createFlowStatus())
 
-        val key1 = FLOW_KEY_1.toString()
-        val key2 = FLOW_KEY_2.toString()
+        val key1 = FlowStatusLookupServiceImplTest.FLOW_KEY_1.toString()
+        val key2 = FlowStatusLookupServiceImplTest.FLOW_KEY_2.toString()
 
         // Persist our first record
-        flowStatusCacheService.initialise(config)
-        flowStatusCacheService.onNext(listOf(record1))
+        flowStatusProcessor.onNext(listOf(record1))
 
         val result1 = stateManager.get(setOf(key1))
 
@@ -227,7 +126,7 @@ class DurableFlowStatusCacheServiceImplTest {
         assertThat(result1[key1]?.version).isEqualTo(0)
 
         // Persist a new record, and update our first
-        flowStatusCacheService.onNext(listOf(record2, record3))
+        flowStatusProcessor.onNext(listOf(record2, record3))
 
         val result2 = stateManager.get(setOf(key1, key2))
